Satisfy Marker Detail module size guardrail
